refactor: add the controller layer to execute a task (#5)
